requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
pyspark>=3.1.2
torch>=1.9.0
stable-baselines3>=1.5.0
gymnasium>=0.28.1
shimmy>=2.0.0

# Performance optimization
numba>=0.53.1
cython>=0.29.24

# Visualization
matplotlib>=3.4.2

# Model optimization
onnx>=1.10.1
onnxruntime>=1.8.1

# Utilities
pyyaml>=5.4.1
